+++ api-admin/controllers/image/upload.js
@@ -11,6 +11,7 @@ import mongoose from 'mongoose';
 import moment from 'moment-timezone';
 import config from 'config';
 import imageServer from 'scp2';
+import md5File from 'md5-file/promise';
 
 import googleCloud from 'google-cloud';
 const gcloud = googleCloud({
@@ -52,6 +53,10 @@ module.exports = async(req, res, next) => {
         let buffer = readChunk.sync(path, 0, 4100);
 
         let { ext } = fileType(buffer);
+
+        // 這段是因為我們要上傳原生圖片廣告給 NOWlink 用，需要知道 md5
+        let md5hash = await md5File(path);
+        console.log(`${originalname} md5 is: ${md5hash}`);
 
         // 用 gm 去讀取圖片的長寬
         let { width, height } = await new Promise((resolve, reject) => {