+++ libs/updateAllHotNews.js
@@ -71,14 +71,15 @@ module.exports = async () => {
                             }
                         },
                         {
-                            $limit : 10
+                            $limit : 15
                         }
                     ]);
                 })
                 .then((pageviews) => {
                     return Promise.mapSeries(pageviews, (pageview) => {
                         return News.findById(pageview._id)
-                            .select('_id sn title shortTitle')
+                            .populate('MainPhoto')
+                            .select('_id sn title shortTitle MainPhoto')
                             .execAsync();
                     });
                 })