+++ pvModels/index.js
@@ -21,15 +21,21 @@ import pageviewSchema from './schemas/pageview';
 import pageviewLogSchema from './schemas/pageviewLog';
 import temperatureLogSchema from './schemas/temperatureLog';
 import personalizeSchema from './schemas/personalize';
+import nowlinkSchema from './schemas/nowlink';
+import nowlinkLogSchema from './schemas/nowlinkLog';
 
 let Pageview = connection.model('Pageview', pageviewSchema);
 let PageviewLog = connection.model('PageviewLog', pageviewLogSchema);
 let TemperatureLog = connection.model('TemperatureLog', temperatureLogSchema);
 let Personalize = connection.model('Personalize', personalizeSchema);
+let Nowlink = connection.model('Nowlink', nowlinkSchema);
+let NowlinkLog = connection.model('NowlinkLog', nowlinkLogSchema);
 
 module.exports = {
     Pageview,
     PageviewLog,
     TemperatureLog,
-    Personalize
+    Personalize,
+    Nowlink,
+    NowlinkLog
 };